resetting migrations & databases and renaming core
