+++ core/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for lymph_interface project.
+Django settings for lymph-interface project.
 
 Generated by 'django-admin startproject' using Django 3.1.6.
 
@@ -133,12 +133,12 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'lymph_interface.urls'
+ROOT_URLCONF = 'core.urls'
 
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [BASE_DIR / "lymph_interface" / "templates"],
+        'DIRS': [BASE_DIR / "core" / "templates"],
         'APP_DIRS': True,
         'OPTIONS': {
             'context_processors': [
@@ -151,7 +151,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'lymph_interface.wsgi.application'
+WSGI_APPLICATION = 'core.wsgi.application'
 
 
 # Database
@@ -206,7 +206,7 @@ USE_TZ = True
 # https://docs.djangoproject.com/en/3.1/howto/static-files/
 
 STATICFILES_DIRS = [
-    BASE_DIR / "lymph_interface" / "static"
+    BASE_DIR / "core" / "static"
 ]
 
 try:
