+++ core/templates/footer.html
