+++ core/templates/head.html
