+++ core/templates/index.md
