+++ core/templates/navbar.html
