+++ core/urls.py
@@ -1,4 +1,4 @@
-"""lymph_interface URL Configuration
+"""lymph-interface URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.1/topics/http/urls/
