--- lymph_interface/static/custom_bulma/sass/usz.scss
@@ -1,23 +1,0 @@
-@charset "utf-8";
-
-// Setting USZ colors
-$usz-blue: #005EA8;
-$usz-green: #00AFA5;
-$usz-orange: #F17900;
-$usz-red: #AE0060;
-$usz-gray: #C5D5DB;
-$light-usz-gray: #dbe6eb;
-$white: #FFFFFF;
-
-// Change Bulma's global variables
-$primary: $usz-blue;
-$primary-dark: $usz-blue;
-$info: $usz-gray;
-$link: $usz-blue;
-$success: $usz-green;
-$warning: $usz-orange;
-$danger: $usz-red;
-
-// importing everything necessary from Bulma
-@import "../node_modules/bulma/bulma.sass";
-@import "../node_modules/@creativebulma/bulma-tooltip/src/sass/index.sass";
